requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
Pillow==7.0.0
pytesseract==0.3.3
requests==2.23.0
urllib3==1.25.8
